LTX2 and LTX2.3 improvements and Tile-size auto-tuning

+++ src/maxdiffusion/models/attention_flax.py
@@ -1914,7 +1914,7 @@ class NNXAttentionOp(nnx.Module):
       self,
       mesh: Mesh,
       attention_kernel: str,
-      scale: int,
+      scale: float,
       heads: int,
       dim_head: int,
       use_memory_efficient_attention: bool = False,
@@ -2009,7 +2009,7 @@ class NNXAttentionOp(nnx.Module):
 class AttentionOp(nn.Module):
   mesh: Mesh
   attention_kernel: str
-  scale: int
+  scale: float
   heads: int
   dim_head: int
   use_memory_efficient_attention: bool = False
